agents/appointments_agent/pyproject.toml
[project]
name = "pet-appointments-agent"
version = "0.1.0"
description = "Pet Clinic Appointments Agent with A2A Protocol"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.5,<0.4.0",
    "httpx>=0.28.1",
    "langchain-openai",
    "langgraph",
    "langchain-core",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "starlette",
    "python-jose[cryptography]>=3.3.0",  # JWT token validation
    "requests>=2.32.0",  # JWKS fetching from Asgardeo
]

[tool.hatch.build.targets.wheel]
packages = ["appoinments"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
